refactor: extract rehearsal song command usecase
